requirements.txt
Django==4.1.6
djangorestframework
djangorestframework-simplejwt
lxml
python-dotenv
psycopg2-binary
djangorestframework-xml
beautifulsoup4